+++ internal/eventhandler/util.go
