--- internal/pkg/assets/.betterglobekey.example.yaml
@@ -1,14 +1,0 @@
-# Double press
-double_press:
-  # Maximum delay between two presses of the Globe key for them to be considered as a double press
-  maximum_delay: 250
-
-# Input sources
-input_sources:
-  # Primary input sources
-  # This is used when the Globe key has been pressed once
-  primary: []
-
-  # Additional input sources
-  # This is used when the Globe key has been double pressed
-  additional: []
